// src/aquanore/enums.d.ts
export enum Keys {
    Escape,
    Up,
    Down,
    Left,
    Right,
    Space,
    Enter,
    LeftShift,
    RightShift,
    LeftCtrl,
    RightCtrl,
    Tab,
    AltLeft,
    AltRight,
    Backspace,
    Home,
    End,
    Insert,
    Delete,
    PageUp,
    PageDown,
    A,
    B,
    C,
    D,
    E,
    F,
    G,
    H,
    I,
    J,
    K,
    L,
    M,
    N,
    O,
    P,
    Q,
    R,
    S,
    T,
    U,
    V,
    W,
    X,
    Y,
    Z,
    D1,
    D2,
    D3,
    D4,
    D5,
    D6,
    D7,
    D8,
    D9,
    D0,
    Minus,
    Equal,
    BracketLeft,
    BracketRight,
    Backslash,
    Slash,
    Comma,
    Period,
    Semicolon,
    Quote,
    Backquote,
    Numpad0,
    Numpad1,
    Numpad2,
    Numpad3,
    Numpad4,
    Numpad5,
    Numpad6,
    Numpad7,
    Numpad8,
    Numpad9,
    NumpadDivide,
    NumpadMultiply,
    NumpadSubtract,
    NumpadAdd,
    NumpadEnter,
    NumpadDecimal,
}

export enum LightType {
    Point,
    Directional
}

export enum SoundStatus {
    Stopped,
    Playing
}